Improve e2e test
Move the test into a describe() block so that the mock constants don't pollute the namespace of the other unit tests.

Second, mock out calls to `Date.now()` so we can better verify that our calls to octokit.checks.create() and octokit.checks.update() use the proper times as parameters. We then are able to some cleaner and more succint jest syntax with toHaveBeenCalledWith() and toHaveBeenCalledTimes().

Third, verify that core.setOutput() was called correctly.

Finally, remove the junky commented out garbage at the top of the file.

TODO might be to better create the Mocktokit class automattically, I bet there's a cleaner way wtih jest.